requirements.txt
cryptography
python-dotenv
telethon
python-telegram-bot
openpyxl
requests
